- Restructured. - Added bake to world with no scale.
