+++ README.md
@@ -1,1 +1,10 @@
 Once Upon A Time, there are crappy tools developed for dreadful Maya.
+
+### Installation
+> Copy ./src/pail to your Maya script directory.
+
+### Cameraman
+```python
+import pail.workbox.cameraman
+pail.workbox.cameraman.show()
+```
